requirements.txt
streamlit
numpy
Pillow
pydicom
opencv-python
plotly
torch
torchvision
sqlalchemy
